Tests/LibWeb/Text/expected/css/scrollable-overflow-recalc-viewport.txt

viewport scrollHeight grew: true
scrollY clamped to grown max: true
relayouts performed: 0
